HEAT-6073857 Added sentinel to prevent exceptions being thrown.
There are cases where $house is not a House, but a boolean (`false`)
(I've been trying to recreate this, but to no avail); if getAddress() is
then called a bit further down the line, this will of course throw an
exception.
This whole defensive programming feels wrong, but been spending too much
time already, and this thing needs to get fixed, not refactored, I'm
sorry.